+++ lib/helper/modular_helper.dart
@@ -139,10 +139,18 @@ abstract class ModularHelper {
     }
   }
 
-  static Future<void> fix({int concurrent = defaultConcurrent}) => execute(
-        ['${FlutterHelper.getCommandDart()} fix --apply'],
-        concurrent: concurrent,
-      );
+  static Future<void> fix([List<String>? paths]) async {
+    if (paths == null || paths.isEmpty) {
+      return execute(['${FlutterHelper.getCommandDart()} fix --apply']);
+    } else {
+      for (var element in paths) {
+        await '${FlutterHelper.getCommandDart()} fix --apply'.start(
+          workingDirectory: element,
+          showLog: false,
+        );
+      }
+    }
+  }
 
   static Future<void> get({int concurrent = defaultConcurrent}) => execute(
         ['${FlutterHelper.getCommandFlutter()} pub get'],